requirements.txt
# python==3.11
openai
numpy
pandas
python-dotenv
pyyaml
tenacity
tiktoken
tqdm
jsonlines
astunparse
pyarrow
google-generativeai
accelerate
matplotlib
